Create Get OS
A simple program to call the OS String from 'settings' table and search for 'CELL-210'.  The CS300 with an internal modem has an OS called CR300-Cell-210.  Parsing out this string can set the logger to communicate in a specific way.  Note while CSBasic is not case sensitive searching in a string is case sensitive